+++ packages/pigeon/lib/src/cpp/cpp_generator.dart
