+++ packages/pigeon/lib/src/dart/dart_generator.dart
