+++ packages/pigeon/lib/src/functional.dart
