+++ packages/pigeon/lib/src/generator.dart
