+++ packages/pigeon/lib/src/generator_tools.dart
