+++ packages/pigeon/lib/src/gobject/gobject_generator.dart
