+++ packages/pigeon/lib/src/java/java_generator.dart
