+++ packages/pigeon/lib/src/swift/templates.dart
@@ -3,7 +3,7 @@
 // found in the LICENSE file.
 
 import '../generator_tools.dart';
-import '../pigeon.dart';
+import '../../pigeon.dart';
 
 /// Name of delegate that handles the callback when an object is deallocated
 /// in an `InstanceManager`.
