+++ packages/pigeon/test/dart_generator_test.dart
@@ -5,9 +5,9 @@
 import 'dart:io' show Directory, File;
 
 import 'package:path/path.dart' as path;
-import 'package:pigeon/ast.dart';
-import 'package:pigeon/dart/dart_generator.dart';
-import 'package:pigeon/generator_tools.dart';
+import 'package:pigeon/src/ast.dart';
+import 'package:pigeon/src/dart/dart_generator.dart';
+import 'package:pigeon/src/generator_tools.dart';
 import 'package:test/test.dart';
 
 const String DEFAULT_PACKAGE_NAME = 'test_package';
