+++ packages/pigeon/test/functional_test.dart
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-import 'package:pigeon/functional.dart';
+import 'package:pigeon/src/functional.dart';
 import 'package:test/test.dart';
 
 void main() {
